refactor: Add Bluebird branding
Starting to replace OpenNMS branding and logos. Redesigning the login page brining Ulf back. Removed duplicated logos
and redesigned the password gate so it tells the user right away what the password complexity rules are instead of
telling him on every mistake.
