+++ features/themes/dashboard-theme/src/main/java/VAADIN/themes/dashboard/styles.css
@@ -50,7 +50,7 @@
 }
 
 #onmslogo {
-    background:url(/opennms/images/o-green-trans.svg) no-repeat;
+    background:url(/opennms/images/bluebird.svg) no-repeat;
     display:block;
     height:44px;
     width:130px;
