+++ opennms-base-assembly/src/main/filtered/etc/opennms.properties
@@ -238,7 +238,7 @@ java.awt.headless=true
 ###### REPORTING ######
 opennms.report.template.dir=${install.dir}/etc
 opennms.report.dir=${install.share.dir}/reports
-opennms.report.logo=${install.webapps.dir}/images/opennms-logo-dark.svg
+opennms.report.logo=${install.webapps.dir}/images/bluebird-logo-dark.svg
 
 # If set to false, scheduled reports are no longer executed/triggered
 opennms.report.scheduler.enabled=true
