+++ ui/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" href="/OpenNMS_Favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>OpenNMS Web Console</title>
+    <title>Bluebird Web Console</title>
   </head>
   <body>
     <div id="app" class="feather-styles"></div>
