+++ ui/src/assets/LogoBluebird.vue
@@ -1,0 +1,18 @@
+  <template>
+    <svg width="100%" height="100%" viewBox="0 0 901 901" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" xml:space="preserve" xmlns:serif="http://www.serif.com/" style="fill-rule:evenodd;clip-rule:evenodd;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:1.5;">
+    <g transform="matrix(1,0,0,1,-509.5,-89.5)">
+        <g>
+            <g transform="matrix(1.02273,0,0,1.02273,100.909,-186.136)">
+                <circle cx="840" cy="710" r="440" style="fill:url(#_Linear1);stroke:black;stroke-width:0.98px;"/>
+            </g>
+          <g transform="matrix(0.999905,0,0,0.999905,150.109,-57.4461)">
+                <path d="M774,680C774,680 780,675 764,674C753.48,673.342 744.667,671.203 735,667C723.5,662 711.888,655.492 700,647C693.218,642.156 682.139,633.435 673,626C658.556,614.25 647.362,603.657 637,592C621,574 608.73,561.459 596,543C579.333,518.833 546.167,468.167 537,449C534.124,442.986 545,437.333 541,432C536,425.333 520.883,416.922 515,413C509,409 499,396 518,403C554.667,417.333 586.667,433 620,443C650.717,452.215 697.468,454.925 724,462C754,470 770,480 770,480C782,484 765,471.18 765,460C765,450 772,442 770,435C768.088,428.308 756.875,426.199 753,420C748,412 750,405.667 752,400C754,394.333 762.833,390.333 765,386C766.789,382.422 766.676,377.632 765,374C763,369.667 756.365,365.048 753,360C749,354 750.229,344.618 753,340C756,335 762.667,333.333 765,330C766.921,327.256 766.776,323.259 766,320C765.167,316.5 762.609,312.261 760,309C756,304 756.876,296.123 760,291C764.167,284.167 780.167,278 785,268C789.833,258 790.153,241.378 789,231C788,222 790.789,211.351 794,206C797,201 804,190 810,200C810,200 832.111,256.149 850,280C870,306.667 911.667,336.667 930,360C943.815,377.583 955.149,398.172 960,420C966.667,450 971.016,486.629 970,510C969,533 965,556 970,560C974.553,563.643 984.341,554.851 990,550C1001.67,540 1016.67,518.333 1030,510C1038.48,504.7 1047,500 1060,500C1072.17,500 1081.67,503.333 1090,510C1097.36,515.89 1104.77,532.155 1110,540C1118,552 1140,560 1140,560L1140,560.861C1140,560.861 1132,568 1120,570C1112.71,571.215 1101.27,574.73 1096,580C1089.33,586.667 1082,597 1080,610C1077.73,624.734 1065.33,655 1057,670C1048.91,684.569 1035.79,698.215 1024,710C1010.67,723.333 991.667,738.333 975,750C959.077,761.146 943.694,765.957 926,774C907.667,782.333 890,785 870,790C850.33,794.918 831,797 810,800C793.174,802.404 776.667,802.667 760,806C743.333,809.333 725,812 710,817C695.858,821.714 685.667,828.333 674,835C663.565,840.963 650,851.667 640,855C630.513,858.162 619.487,856.838 610,860C600,863.333 588.556,873.698 577,877C565.333,880.333 550,879.333 540,876C531.056,873.019 526.667,864.667 520,863C513.333,861.333 506.667,870 500,870C493.333,870 480,860 480,860C478,837 524.154,821.69 540,814C562.667,803 594.333,792.333 620,780C641.774,769.537 665.667,757 684,746C700.017,736.39 714.88,724.966 730,714C745.167,703 774,680 774,680L730,714L684,746" style="fill:rgb(15,17,42);stroke:black;stroke-width:1px;"/>
+            </g>
+        </g>
+    </g>
+      <defs>
+        <linearGradient id="_Linear1" x1="0" y1="0" x2="1" y2="0" gradientUnits="userSpaceOnUse" gradientTransform="matrix(695.423,-537.778,537.778,695.423,480.926,964.222)"><stop offset="0" style="stop-color:rgb(228,97,255);stop-opacity:1"/><stop offset="0.43" style="stop-color:rgb(158,105,255);stop-opacity:1"/><stop offset="1" style="stop-color:rgb(0,123,255);stop-opacity:1"/></linearGradient>
+    </defs>
+</svg>
+</template>
+<script lang="ts" setup></script>
