--- ui/src/assets/LogoHorizon.vue
@@ -1,103 +1,0 @@
-<template>
-    <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" viewBox="0 0 624.09 69.3">
-        <defs>
-            <linearGradient id="linear-gradient" x1="38.51" y1="67.61" x2="-3.17" y2="25.53"
-                gradientUnits="userSpaceOnUse">
-                <stop offset="0" stop-color="#0081ad" />
-                <stop offset="1" stop-color="#14d1df" />
-            </linearGradient>
-            <linearGradient id="linear-gradient-2" x1="38.62" y1="67.5" x2="-3.06" y2="25.41"
-                xlink:href="#linear-gradient" />
-            <linearGradient id="linear-gradient-3" x1="25.87" y1="-3" x2="65.55" y2="37.08"
-                xlink:href="#linear-gradient" />
-            <linearGradient id="linear-gradient-4" x1="25.98" y1="-3.12" x2="65.67" y2="36.97"
-                xlink:href="#linear-gradient" />
-            <linearGradient id="linear-gradient-5" x1="8.32" y1="8.72" x2="53.79" y2="53.6"
-                gradientUnits="userSpaceOnUse">
-                <stop offset="0" stop-color="#14d1df" />
-                <stop offset="1" stop-color="#85d9a5" />
-            </linearGradient>
-            <linearGradient id="linear-gradient-6" x1="8.33" y1="8.71" x2="53.81" y2="53.58"
-                xlink:href="#linear-gradient-5" />
-        </defs>
-        <g id="Layer_2" data-name="Layer 2">
-            <g id="Lockups">
-                <path class="cls-1"
-                    d="M247.71,11.78h5.56L278.71,48H279V11.78h5.49V57.15H279L253.53,21h-.32V57.15h-5.5Z" />
-                <path class="cls-1"
-                    d="M292.18,11.78h6.13l17.58,38.28,17.63-38.28h6.14V57.15h-5.5V22.71h-.32L318,57.15h-4.16L298,22.71h-.32V57.15h-5.49Z" />
-                <path class="cls-1"
-                    d="M345.48,46.74,350,44.63a10.66,10.66,0,0,0,10.67,8c4.79,0,8.95-2.62,8.95-7.8,0-4.66-3.33-6.71-8.63-8.81l-3-1.15c-6.65-2.69-10.68-6.4-10.68-12.27,0-7,5.69-11.51,13-11.51a14,14,0,0,1,13.68,9.46l-4.22,2.17c-1.79-4.09-4.73-6.58-9.33-6.58-4.35,0-7.67,2.43-7.67,6.52,0,3.19,1.92,5.69,7.29,7.8l3.06,1.21C370.6,34.66,375.26,38,375.26,45c0,8.37-7,12.84-14.57,12.84S347,53.26,345.48,46.74Z" />
-                <path class="cls-1"
-                    d="M78.54,34.47c0-13.87,10.67-23.33,23.51-23.33s23.52,9.46,23.52,23.33-10.74,23.32-23.52,23.32S78.54,48.34,78.54,34.47Zm36.61,0c0-8.37-5.94-13.74-13.1-13.74S89,26.1,89,34.47s6,13.74,13.1,13.74S115.15,42.84,115.15,34.47Z" />
-                <path class="cls-1"
-                    d="M167,40.54c0,11.57-8.3,17.25-15.85,17.25-5.87,0-9.33-3.19-9.84-4.41h-.19V69.3h-9.9V23.92h9.9v3.77h.19c.51-1.21,4-4.41,9.84-4.41C158.7,23.28,167,29,167,40.54Zm-9.71,0a8.21,8.21,0,1,0-16.42,0,8.21,8.21,0,1,0,16.42,0Z" />
-                <path class="cls-1"
-                    d="M203,44.37H179.39c1.34,3.58,4.66,5.37,9.07,5.37a12.59,12.59,0,0,0,8-2.55l4.48,4.66c-2.75,3.45-7.48,5.94-13.55,5.94-9.21,0-17.51-6.26-17.51-17.19,0-11.24,8.37-17.32,16.93-17.32S205.14,29.16,203,44.37ZM179.2,37.28h15.14c-.7-3.83-4.09-5.69-7.41-5.69A7.64,7.64,0,0,0,179.2,37.28Z" />
-                <path class="cls-1"
-                    d="M240.06,37.6V57.15h-9.91V39.32c0-4.15-2.11-7.22-6.2-7.22s-6.13,3.07-6.13,7.22V57.15h-9.91V23.92h9.91v4.22c1.85-2.81,5.11-4.86,9.84-4.86C234.37,23.28,240.06,27.44,240.06,37.6Z" />
-                <path class="cls-2"
-                    d="M17.61,36a103.5,103.5,0,0,1-11-14,42.82,42.82,0,0,1-2.92-5.32A31.63,31.63,0,0,0,7.89,52.51C8.26,49.1,11.89,43,17.61,36Z" />
-                <path class="cls-3"
-                    d="M27.2,45.61c-7,5.73-13.07,9.36-16.49,9.72a31.65,31.65,0,0,0,35.83,4.16A49.57,49.57,0,0,1,38.83,55,114.14,114.14,0,0,1,27.2,45.61Z" />
-                <path class="cls-4"
-                    d="M36,17.61c7-5.72,13.08-9.35,16.49-9.72A31.63,31.63,0,0,0,16.68,3.74,49.16,49.16,0,0,1,24.4,8.23,111.51,111.51,0,0,1,36,17.61Z" />
-                <path class="cls-5"
-                    d="M55.33,10.71c-.36,3.42-4,9.53-9.72,16.49a105.09,105.09,0,0,1,11,14,43.44,43.44,0,0,1,2.91,5.33A31.65,31.65,0,0,0,55.33,10.71Z" />
-                <path class="cls-6"
-                    d="M31.61,21.42h0C20.75,11.6,10.93,6,8.44,8.44c-2.88,2.89,5.15,15.6,18,28.4,1.75,1.75,3.5,3.41,5.22,5Z" />
-                <path class="cls-7"
-                    d="M41.81,31.61h0c-1.56,1.73-3.22,3.48-5,5.23s-3.5,3.41-5.23,5c10.87,9.82,20.68,15.46,23.17,13S51.63,42.48,41.81,31.61Z" />
-                <path class="cls-8" d="M436.8,11.78V57.15H428V38.56H409.38V57.15h-8.75V11.78h8.75V30.51H428V11.78Z" />
-                <path class="cls-8"
-                    d="M441.71,40.54c0-11.25,8.82-17.26,17.51-17.26s17.51,6,17.51,17.26-8.88,17.25-17.51,17.25S441.71,51.79,441.71,40.54Zm26.52,0c0-5.94-4.21-9.52-9-9.52s-9.07,3.58-9.07,9.52,4.28,9.52,9.07,9.52S468.23,46.48,468.23,40.54Z" />
-                <path class="cls-8"
-                    d="M502.87,23.54v8.18a12.32,12.32,0,0,0-2.17-.19c-6.46,0-10.49,4.47-10.49,11.12v14.5h-8.56V23.92h8.56v8.63c1.16-5.11,4.41-9.27,10.42-9.27A9,9,0,0,1,502.87,23.54Z" />
-                <path class="cls-8" d="M507.72,11.78h8.57v7.54h-8.57Zm0,12.14h8.57V57.15h-8.57Z" />
-                <path class="cls-8" d="M538.33,31.4H522.42V23.92h29.65L534.31,49.61h16.48v7.54H520.57Z" />
-                <path class="cls-8"
-                    d="M552.84,40.54c0-11.25,8.81-17.26,17.51-17.26s17.51,6,17.51,17.26S579,57.79,570.35,57.79,552.84,51.79,552.84,40.54Zm26.52,0c0-5.94-4.22-9.52-9-9.52s-9.08,3.58-9.08,9.52,4.28,9.52,9.08,9.52S579.36,46.48,579.36,40.54Z" />
-                <path class="cls-8"
-                    d="M624.09,37.6V57.15h-8.57V39c0-4.86-2.62-8.06-7.09-8.06s-7.09,3.2-7.09,8.06V57.15h-8.57V23.92h8.57v4.54c1.85-2.94,5.17-5.18,10.16-5.18C618.33,23.28,624.09,27.5,624.09,37.6Z" />
-            </g>
-        </g>
-    </svg>
-</template>
-<script lang="ts" setup>
-</script>
-<style lang="scss" scoped>
-svg {
-    max-width:198px;
-}
-.cls-1 {
-    fill: #fff;
-}
-
-.cls-2 {
-    fill: url(#linear-gradient);
-}
-
-.cls-3 {
-    fill: url(#linear-gradient-2);
-}
-
-.cls-4 {
-    fill: url(#linear-gradient-3);
-}
-
-.cls-5 {
-    fill: url(#linear-gradient-4);
-}
-
-.cls-6 {
-    fill: url(#linear-gradient-5);
-}
-
-.cls-7 {
-    fill: url(#linear-gradient-6);
-}
-
-.cls-8 {
-    fill: #0081ad;
-}
-</style>
