+++ ui/src/components/Layout/Menubar.vue
@@ -256,7 +256,7 @@ import ArrowDropDown from '@featherds/icon/navigation/ArrowDropDown'
 import LightDarkMode from '@featherds/icon/action/LightDarkMode'
 import UpdateUtilities from '@featherds/icon/action/UpdateUtilities'
 import Person from '@featherds/icon/action/Person'
-import Logo from '@/assets/LogoHorizon.vue'
+import Logo from '@/assets/LogoBluebird.vue'
 import { useAppStore } from '@/stores/appStore'
 import { useMenuStore } from '@/stores/menuStore'
 import { usePluginStore } from '@/stores/pluginStore'
